+++ relay-general/src/processor/selector.rs
@@ -14,6 +14,9 @@ pub enum InvalidSelectorError {
     #[fail(display = "invalid selector: deep wildcard used more than once")]
     InvalidDeepWildcard,
 
+    #[fail(display = "invalid selector: wildcard must be part of a path")]
+    InvalidWildcard,
+
     #[fail(display = "invalid selector: {}", _0)]
     ParseError(Error<Rule>),
 
@@ -255,7 +258,7 @@ fn handle_selector(pair: Pair<Rule>) -> Result<SelectorSpec, InvalidSelectorErro
         }
         Rule::SelectorPath => {
             let mut used_deep_wildcard = false;
-            let items = pair
+            let items: Vec<SelectorPathItem> = pair
                 .into_inner()
                 .map(|item| {
                     let rv = handle_selector_path_item(item)?;
@@ -269,6 +272,10 @@ fn handle_selector(pair: Pair<Rule>) -> Result<SelectorSpec, InvalidSelectorErro
                     Ok(rv)
                 })
                 .collect::<Result<_, _>>()?;
+
+            if matches!(items.as_slice(), [SelectorPathItem::Wildcard]) {
+                return Err(InvalidSelectorError::InvalidWildcard);
+            }
 
             Ok(SelectorSpec::Path(items))
         }
@@ -340,4 +347,16 @@ fn test_roundtrip() {
     check_roundtrip("!a && !b");
     check_roundtrip("!(a && !b)");
     check_roundtrip("!(a && b)");
+}
+
+#[test]
+fn test_invalid() {
+    assert!(matches!(
+        SelectorSpec::from_str("* && foo"),
+        Err(InvalidSelectorError::InvalidWildcard)
+    ));
+    assert!(matches!(
+        SelectorSpec::from_str("$frame.**.foo.**"),
+        Err(InvalidSelectorError::InvalidDeepWildcard)
+    ));
 }